requirement.txt
numpy==1.21.5
Pillow==9.0.0